Move APIs to work with health checks into translator package to prepare for future refactoring
